[INLONG-7926][TubeMQ] Refactor the interface design under the client package
